# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Um repositorio de autores e suas obras",
        "description": "Um repositorio para agrupar todos os autores e suas obras conhecidas da humanidade",
        "termsOfService": "https://minhaapi.com/terms",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/"
        }
    ],
    "paths": {
        "/Login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login do usuario",
                "description": "Retorna o token jwt para o login do usuario",
                "operationId": "7891ab35e16767f4808b265e4a3cc62c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta de sucesso: ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "string": {
                                            "description": " string indicando o token",
                                            "format": "string",
                                            "default": "token de acesso : "
                                        },
                                        "token": {
                                            "description": "token",
                                            "format": "string",
                                            "default": "20d338931e8d6bd9466edeba78ea7dce7c7bc01aa5cc5b4735691c50a2fe3228"
                                        },
                                        "integer": {
                                            "description": "status da resposta",
                                            "format": "integer",
                                            "default": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout do usuario",
                "description": "Faz o logout do usuario",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Resposta de sucesso: ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "string": {
                                            "description": " string indicando a mensagem",
                                            "format": "string",
                                            "default": " msg"
                                        },
                                        "": {
                                            "description": "mensagem falando que o usuario foi deslogado",
                                            "format": "string",
                                            "default": "logout realizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh do token",
                "description": "Faz o refresh do token",
                "operationId": "0c3ea3c61aed881016646810460499e9",
                "responses": {
                    "200": {
                        "description": "Resposta de sucesso: ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "string": {
                                            "description": " mensagem indicando o token",
                                            "format": "string",
                                            "default": " token"
                                        },
                                        "token": {
                                            "description": "token",
                                            "format": "string",
                                            "default": "20d338931e8d6bd9466edeba78ea7dce7c7bc01aa5cc5b4735691c50a2fe3228"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Pegue o token indo na operação login e passando o email e senha",
                "name": "Token JWT",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}